# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Energy and Water",
        "version": "1.0.0"
    },
    "paths": {
        "/v1/categoria_servicio": {
            "get": {
                "summary": "Obtener todas las categorías de servicio",
                "operationId": "getAllCategoriasServicio",
                "responses": {
                    "200": {
                        "description": "Operación exitosa"
                    }
                }
            },
            "post": {
                "summary": "Agregar una nueva categoría de servicio",
                "operationId": "addCategoriaServicio",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombre",
                                    "imagen",
                                    "texto",
                                    "activo"
                                ],
                                "properties": {
                                    "nombre": {
                                        "type": "string"
                                    },
                                    "imagen": {
                                        "type": "string"
                                    },
                                    "texto": {
                                        "type": "string"
                                    },
                                    "activo": {
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Información de Categoría de Servicio agregada con éxito"
                    },
                    "400": {
                        "description": "Validación fallida"
                    }
                }
            }
        },
        "/v1/categoria_servicio/{id}": {
            "get": {
                "summary": "Obtener una categoría de servicio por ID",
                "operationId": "getCategoriaServicioById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la categoría de servicio",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operación exitosa"
                    },
                    "404": {
                        "description": "No se encontraron datos"
                    }
                }
            },
            "put": {
                "summary": "Actualizar o crear una categoría de servicio",
                "operationId": "updateOrCreateCategoriaServicio",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la categoría de servicio",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombre",
                                    "imagen",
                                    "texto",
                                    "activo"
                                ],
                                "properties": {
                                    "nombre": {
                                        "type": "string"
                                    },
                                    "imagen": {
                                        "type": "string"
                                    },
                                    "texto": {
                                        "type": "string"
                                    },
                                    "activo": {
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Información de categoria_servicio actualizada con éxito"
                    },
                    "201": {
                        "description": "Información de categoria_servicio agregada con éxito"
                    },
                    "409": {
                        "description": "Ya existe un registro con los datos proporcionados"
                    },
                    "500": {
                        "description": "Error al procesar la solicitud"
                    }
                }
            },
            "delete": {
                "summary": "Eliminar una categoría de servicio por ID",
                "operationId": "deleteCategoriaServicioById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la categoría de servicio",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Información de categoria_servicio eliminada con éxito"
                    },
                    "404": {
                        "description": "No se encontró el id"
                    }
                }
            },
            "patch": {
                "summary": "Actualizar parcialmente una categoría de servicio",
                "operationId": "partialUpdateCategoriaServicio",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la categoría de servicio",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nombre": {
                                        "type": "string"
                                    },
                                    "imagen": {
                                        "type": "string"
                                    },
                                    "texto": {
                                        "type": "string"
                                    },
                                    "activo": {
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Información de categoria servicio actualizada con éxito"
                    },
                    "400": {
                        "description": "No hay datos para actualizar"
                    },
                    "404": {
                        "description": "No se encontró el id"
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                }
            }
        },
...
    }
}